Rename `entrypoint.module` to `entrypoint.api`
