+++ src/lib/getSourceBranchFromCommits.ts
@@ -1,4 +1,4 @@
-import { Commit } from '../entrypoint.module';
+import { Commit } from '../entrypoint.api';
 
 export function getSourceBranchFromCommits(commits: Commit[]) {
   // sourceBranch should be the same for all commits, so picking `sourceBranch` from the first commit should be fine 🤞
