+++ src/test/e2e/module/entrypoint.module.mutation.test.ts
@@ -1,5 +1,5 @@
 import { Octokit } from '@octokit/rest';
-import { BackportResponse, backportRun } from '../../../entrypoint.module';
+import { BackportResponse, backportRun } from '../../../entrypoint.api';
 import { getShortSha } from '../../../lib/github/commitFormatters';
 import { getDevAccessToken } from '../../private/getDevAccessToken';
 import { getSandboxPath, resetSandbox } from '../../sandbox';
