+++ src/test/handleUnbackportedPullRequests.private.test.ts
@@ -1,4 +1,4 @@
-import { getCommits, backportRun, Commit } from '../entrypoint.module';
+import { getCommits, backportRun, Commit } from '../entrypoint.api';
 import { getDevAccessToken } from './private/getDevAccessToken';
 import { getSandboxPath, resetSandbox } from './sandbox';
