+++ src/utils/filterUnmergedCommits.ts
@@ -1,4 +1,4 @@
-import { Commit } from '../entrypoint.module';
+import { Commit } from '../entrypoint.api';
 
 export function filterUnmergedCommits(commit: Commit) {
   return commit.targetPullRequestStates.some((pr) => pr.state !== 'MERGED');
